refactor(session-store): extract transcript-delta + token-stream helpers (file #4)
Lift applyTranscriptDeltaToSnapshot + row-token-stream map utils + appended-word
counting into transcript-delta.ts (reuses transcript-entry-index); session-store
re-exports the public fns. Net 13/13, store vitest parity (503 pass), check 0.
